requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
openpyxl==3.1.2
fpdf==1.7.2
Pillow
python-dotenv==1.0.0
email-validator==1.4.0
gunicorn==20.1.0
PyMySQL==1.1.0
cryptography==41.0.7 